Add justfile and put input tests seperate
