+++ src/utils/tests.rs
@@ -426,210 +426,3 @@ fn draw_help_message() {
     let res = String::from_utf8(out).expect("Should have written valid UTF-8");
     assert!(res.contains("minus"));
 }
-
-#[test]
-#[allow(clippy::too_many_lines)]
-fn input_handling() {
-    let mut pager = Pager::new();
-    pager.upper_mark = 12;
-    pager.set_line_numbers(LineNumbers::Enabled);
-    pager.rows = 5;
-    pager.running = true;
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Down,
-            modifiers: KeyModifiers::NONE,
-        });
-        assert_eq!(
-            Some(InputEvent::UpdateUpperMark(pager.upper_mark + 1)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Up,
-            modifiers: KeyModifiers::NONE,
-        });
-        assert_eq!(
-            Some(InputEvent::UpdateUpperMark(pager.upper_mark - 1)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Down,
-            modifiers: KeyModifiers::NONE,
-        });
-        let mut pager = pager.clone();
-        pager.upper_mark = usize::MAX;
-        assert_eq!(
-            Some(InputEvent::UpdateUpperMark(usize::MAX)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Up,
-            modifiers: KeyModifiers::NONE,
-        });
-        let mut pager = pager.clone();
-        pager.upper_mark = usize::MIN;
-        assert_eq!(
-            Some(InputEvent::UpdateUpperMark(usize::MIN)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Mouse(MouseEvent::ScrollDown(0, 0, KeyModifiers::NONE));
-        assert_eq!(
-            Some(InputEvent::UpdateUpperMark(pager.upper_mark + 5)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Mouse(MouseEvent::ScrollUp(0, 0, KeyModifiers::NONE));
-        assert_eq!(
-            Some(InputEvent::UpdateUpperMark(pager.upper_mark - 5)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Char('g'),
-            modifiers: KeyModifiers::NONE,
-        });
-        assert_eq!(
-            Some(InputEvent::UpdateUpperMark(0)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::PageUp,
-            modifiers: KeyModifiers::NONE,
-        });
-        assert_eq!(
-            // rows is 5, therefore upper_mark = upper_mark - rows -1
-            Some(InputEvent::UpdateUpperMark(8)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Char('g'),
-            modifiers: KeyModifiers::SHIFT,
-        });
-        assert_eq!(
-            Some(InputEvent::UpdateUpperMark(usize::MAX)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Char('G'),
-            modifiers: KeyModifiers::NONE,
-        });
-        assert_eq!(
-            Some(InputEvent::UpdateUpperMark(usize::MAX)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Char('G'),
-            modifiers: KeyModifiers::SHIFT,
-        });
-        assert_eq!(
-            Some(InputEvent::UpdateUpperMark(usize::MAX)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::PageDown,
-            modifiers: KeyModifiers::NONE,
-        });
-        assert_eq!(
-            // rows is 5, therefore upper_mark = upper_mark - rows -1
-            Some(InputEvent::UpdateUpperMark(16)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Resize(42, 35);
-        assert_eq!(
-            Some(InputEvent::UpdateTermArea(42, 35)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Char('l'),
-            modifiers: KeyModifiers::CONTROL,
-        });
-        assert_eq!(
-            Some(InputEvent::UpdateLineNumber(!pager.line_numbers)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Char('q'),
-            modifiers: KeyModifiers::NONE,
-        });
-        assert_eq!(Some(InputEvent::Exit), handle_input(ev, &pager));
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Char('c'),
-            modifiers: KeyModifiers::CONTROL,
-        });
-        assert_eq!(Some(InputEvent::Exit), handle_input(ev, &pager));
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Char('a'),
-            modifiers: KeyModifiers::NONE,
-        });
-        assert_eq!(None, handle_input(ev, &pager));
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Char('/'),
-            modifiers: KeyModifiers::NONE,
-        });
-        assert_eq!(
-            Some(InputEvent::Search(SearchMode::Forward)),
-            handle_input(ev, &pager)
-        );
-    }
-
-    {
-        let ev = Event::Key(KeyEvent {
-            code: KeyCode::Char('?'),
-            modifiers: KeyModifiers::NONE,
-        });
-        assert_eq!(
-            Some(InputEvent::Search(SearchMode::Reverse)),
-            handle_input(ev, &pager)
-        );
-    }
-}
